+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_file_client.hpp
@@ -8,7 +8,7 @@
 #include <string>
 
 #include <azure/core/credentials.hpp>
-#include <azure/core/http/pipeline.hpp>
+#include <azure/core/internal/http/pipeline.hpp>
 #include <azure/core/response.hpp>
 #include <azure/storage/blobs/block_blob_client.hpp>
 #include <azure/storage/common/storage_credential.hpp>
@@ -273,7 +273,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
         Azure::Core::Http::Url fileUrl,
         Blobs::BlobClient blobClient,
         Blobs::BlockBlobClient blockBlobClient,
-        std::shared_ptr<Azure::Core::Http::HttpPipeline> pipeline)
+        std::shared_ptr<Azure::Core::Internal::Http::HttpPipeline> pipeline)
         : DataLakePathClient(std::move(fileUrl), std::move(blobClient), pipeline),
           m_blockBlobClient(std::move(blockBlobClient))
     {