+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/protocol/share_rest_client.hpp
@@ -18,7 +18,7 @@
 #include <azure/core/datetime.hpp>
 #include <azure/core/etag.hpp>
 #include <azure/core/http/http.hpp>
-#include <azure/core/http/pipeline.hpp>
+#include <azure/core/internal/http/pipeline.hpp>
 #include <azure/core/internal/json.hpp>
 #include <azure/core/nullable.hpp>
 #include <azure/core/response.hpp>
@@ -1115,7 +1115,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ServiceSetPropertiesResult> SetProperties(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const SetPropertiesOptions& setPropertiesOptions)
         {
@@ -1152,7 +1152,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ServiceGetPropertiesResult> GetProperties(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const GetPropertiesOptions& getPropertiesOptions)
         {
@@ -1182,7 +1182,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ServiceListSharesSinglePageResult> ListSharesSinglePage(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const ListSharesSinglePageOptions& listSharesSinglePageOptions)
         {
@@ -2544,7 +2544,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareCreateResult> Create(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const CreateOptions& createOptions)
         {
@@ -2586,7 +2586,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareGetPropertiesResult> GetProperties(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const GetPropertiesOptions& getPropertiesOptions)
         {
@@ -2626,7 +2626,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareDeleteResult> Delete(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const DeleteOptions& deleteOptions)
         {
@@ -2670,7 +2670,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareAcquireLeaseResult> AcquireLease(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const AcquireLeaseOptions& acquireLeaseOptions)
         {
@@ -2715,7 +2715,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareReleaseLeaseResult> ReleaseLease(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const ReleaseLeaseOptions& releaseLeaseOptions)
         {
@@ -2754,7 +2754,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareChangeLeaseResult> ChangeLease(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const ChangeLeaseOptions& changeLeaseOptions)
         {
@@ -2798,7 +2798,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareRenewLeaseResult> RenewLease(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const RenewLeaseOptions& renewLeaseOptions)
         {
@@ -2837,7 +2837,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareBreakLeaseResult> BreakLease(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const BreakLeaseOptions& breakLeaseOptions)
         {
@@ -2883,7 +2883,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareCreateSnapshotResult> CreateSnapshot(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const CreateSnapshotOptions& createSnapshotOptions)
         {
@@ -2915,7 +2915,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareCreatePermissionResult> CreatePermission(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const CreatePermissionOptions& createPermissionOptions)
         {
@@ -2952,7 +2952,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareGetPermissionResult> GetPermission(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const GetPermissionOptions& getPermissionOptions)
         {
@@ -2983,7 +2983,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareSetPropertiesResult> SetProperties(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const SetPropertiesOptions& setPropertiesOptions)
         {
@@ -3027,7 +3027,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareSetMetadataResult> SetMetadata(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const SetMetadataOptions& setMetadataOptions)
         {
@@ -3064,7 +3064,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareGetAccessPolicyResult> GetAccessPolicy(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const GetAccessPolicyOptions& getAccessPolicyOptions)
         {
@@ -3097,7 +3097,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareSetAccessPolicyResult> SetAccessPolicy(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const SetAccessPolicyOptions& setAccessPolicyOptions)
         {
@@ -3140,7 +3140,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareGetStatisticsResult> GetStatistics(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const GetStatisticsOptions& getStatisticsOptions)
         {
@@ -3173,7 +3173,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<ShareRestoreResult> Restore(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const RestoreOptions& restoreOptions)
         {
@@ -4047,7 +4047,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<DirectoryCreateResult> Create(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const CreateOptions& createOptions)
         {
@@ -4091,7 +4091,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<DirectoryGetPropertiesResult> GetProperties(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const GetPropertiesOptions& getPropertiesOptions)
         {
@@ -4123,7 +4123,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<DirectoryDeleteResult> Delete(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const DeleteOptions& deleteOptions)
         {
@@ -4153,7 +4153,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<DirectorySetPropertiesResult> SetProperties(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const SetPropertiesOptions& setPropertiesOptions)
         {
@@ -4196,7 +4196,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<DirectorySetMetadataResult> SetMetadata(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const SetMetadataOptions& setMetadataOptions)
         {
@@ -4232,7 +4232,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
         static Azure::Core::Response<DirectoryListFilesAndDirectoriesSinglePageResult>
         ListFilesAndDirectoriesSinglePage(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const ListFilesAndDirectoriesSinglePageOptions&
                 listFilesAndDirectoriesSinglePageOptions)
@@ -4293,7 +4293,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<DirectoryListHandlesResult> ListHandles(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const ListHandlesOptions& listHandlesOptions)
         {
@@ -4349,7 +4349,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<DirectoryForceCloseHandlesResult> ForceCloseHandles(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const ForceCloseHandlesOptions& forceCloseHandlesOptions)
         {
@@ -5234,7 +5234,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileCreateResult> Create(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const CreateOptions& createOptions)
         {
@@ -5318,7 +5318,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileDownloadResult> Download(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const DownloadOptions& downloadOptions)
         {
@@ -5358,7 +5358,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileGetPropertiesResult> GetProperties(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const GetPropertiesOptions& getPropertiesOptions)
         {
@@ -5395,7 +5395,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileDeleteResult> Delete(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const DeleteOptions& deleteOptions)
         {
@@ -5436,7 +5436,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileSetHttpHeadersResult> SetHttpHeaders(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const SetHttpHeadersOptions& setHttpHeadersOptions)
         {
@@ -5524,7 +5524,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileSetMetadataResult> SetMetadata(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const SetMetadataOptions& setMetadataOptions)
         {
@@ -5561,7 +5561,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileAcquireLeaseResult> AcquireLease(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const AcquireLeaseOptions& acquireLeaseOptions)
         {
@@ -5597,7 +5597,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileReleaseLeaseResult> ReleaseLease(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const ReleaseLeaseOptions& releaseLeaseOptions)
         {
@@ -5627,7 +5627,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileChangeLeaseResult> ChangeLease(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const ChangeLeaseOptions& changeLeaseOptions)
         {
@@ -5662,7 +5662,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileBreakLeaseResult> BreakLease(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const BreakLeaseOptions& breakLeaseOptions)
         {
@@ -5699,7 +5699,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
         static Azure::Core::Response<FileUploadRangeResult> UploadRange(
             const Azure::Core::Http::Url& url,
             Azure::Core::Http::BodyStream& bodyStream,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const UploadRangeOptions& uploadRangeOptions)
         {
@@ -5748,7 +5748,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileUploadRangeFromUrlResult> UploadRangeFromUrl(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const UploadRangeFromUrlOptions& uploadRangeFromUrlOptions)
         {
@@ -5816,7 +5816,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileGetRangeListResult> GetRangeList(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const GetRangeListOptions& getRangeListOptions)
         {
@@ -5875,7 +5875,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileStartCopyResult> StartCopy(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const StartCopyOptions& startCopyOptions)
         {
@@ -5955,7 +5955,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileAbortCopyResult> AbortCopy(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const AbortCopyOptions& abortCopyOptions)
         {
@@ -5992,7 +5992,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileListHandlesResult> ListHandles(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const ListHandlesOptions& listHandlesOptions)
         {
@@ -6041,7 +6041,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
         static Azure::Core::Response<FileForceCloseHandlesResult> ForceCloseHandles(
             const Azure::Core::Http::Url& url,
-            Azure::Core::Http::HttpPipeline& pipeline,
+            Azure::Core::Internal::Http::HttpPipeline& pipeline,
             Azure::Core::Context context,
             const ForceCloseHandlesOptions& forceCloseHandlesOptions)
         {