+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_service_client.hpp
@@ -6,7 +6,7 @@
 #include <memory>
 #include <string>
 
-#include <azure/core/http/pipeline.hpp>
+#include <azure/core/internal/http/pipeline.hpp>
 #include <azure/core/response.hpp>
 #include <azure/storage/common/storage_credential.hpp>
 
@@ -104,6 +104,6 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
   private:
     Azure::Core::Http::Url m_serviceUrl;
-    std::shared_ptr<Azure::Core::Http::HttpPipeline> m_pipeline;
+    std::shared_ptr<Azure::Core::Internal::Http::HttpPipeline> m_pipeline;
   };
 }}}} // namespace Azure::Storage::Files::Shares